requirements.txt
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
crcmod==1.7
cryptography==43.0.3
cycler==0.12.1
filelock==3.16.1
Flask==3.0.3
Flask-Cors==5.0.0
fonttools==4.54.1
fsspec==2024.9.0
google
google-api-core==2.23.0
google-api-python-client
google-auth==2.36.0
google-auth-httplib2
google-auth-oauthlib
google-cloud-core==2.4.1
google-cloud-secret-manager
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-pasta
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grpc-google-iam-v1
huggingface-hub==0.25.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==0.10.0
json5==0.9.25
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.7
logging
MarkupSafe==3.0.1
matplotlib==3.9.2
mpmath==1.3.0
networkx==3.4
numpy==1.23.5
opencv-python==4.10.0.84
packaging==24.1
pandas==2.0.0
pillow==11.0.0
proto-plus==1.25.0
protobuf==5.28.3
psutil==6.0.0
py-cpuinfo==9.0.0
pyarrow==14.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.21.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
pytz==2024.2
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
rsa==4.9
safetensors==0.4.5
scipy==1.14.1
seaborn==0.13.2
six==1.16.0
sympy==1.13.3
thop==0.1.1.post2209072238
tokenizers==0.20.1
torch==2.4.1
torchvision==0.19.1
tqdm==4.66.5
transformers==4.45.2
typing_extensions==4.12.2
tzdata==2024.2
ultralytics==8.2.0
urllib3==2.2.3
Werkzeug==3.0.4
